Use `re_path` for Ask CFPB export URL
This change uses `re_path` for the Ask CFPB Export URL definition.

`re_path` and `url` should be interchangable because they both take a regular expression. `path` takes a slightly different non-regex string. When it doesn't have a trailing slash `'export-ask/` it breaks the expectation that the URL ends with a `/`.

I've chosen to revert to using a regular expression for maximum compatibility — this way there's no unexpected difference between Django 1.11 and 2.0.